pyproject.toml
[project]
name = "food-safety-detection"
version = "0.1.0"
description = "Automated food safety event detection from news articles using LLMs"
authors = [
    {name = "Lowren Hwa", email = "lowren@example.com"}
]
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    # Deep Learning
    "torch>=2.0.0",
    "transformers>=4.36.0",
    "peft>=0.7.0",
    "accelerate>=0.25.0",
    "trl>=0.7.0",
    # PyTorch Lightning
    "pytorch-lightning>=2.0.0",
    "torchmetrics>=1.0.0",
    # Data Science
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "xgboost>=2.0.0",
    "imbalanced-learn>=0.11.0",
    # NLP
    "nltk>=3.8",
    "sentence-transformers>=2.2.0",
    "setfit>=1.0.0",
    "spacy>=3.7.0",
    # Visualization
    "plotly>=5.14.0",
    "seaborn>=0.12.0",
    "matplotlib>=3.7.0",
    "wordcloud>=1.9.0",
    "kaleido>=0.2.1",
    # Configuration
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "python-dotenv>=1.0.0",
    # Experiment Tracking
    "wandb>=0.15.0",
    "mlflow>=2.8.0",
    # Utilities
    "tqdm>=4.65.0",
    "rich>=13.0.0",
    "python-json-logger>=2.0.0",
    "jupyterlab>=4.0.0",
    "ipywidgets>=8.1.0",
    "textstat>=0.7.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.5.0",
]

unsloth = []

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py310']
include = '\.pyi?$'

[tool.ruff]
line-length = 100
target-version = "py310"
select = ["E", "F", "I", "N", "W", "B", "C4", "UP"]
ignore = ["E501", "B008", "B905"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
